+++ QUAD/gsas_tools.py
@@ -74,6 +74,7 @@ class Calculator:
         G2path.SetBinaryPath()
         varyList = []
         parmDict = {}
+        symmetry = {}
         Controls = G2stIO.GetControls(GPXfile)
         calcControls = {}
         calcControls.update(Controls)
@@ -104,6 +105,12 @@ class Calculator:
         parmDict.update(hapDict)
         parmDict.update(histDict)
         G2stIO.GetFprime(calcControls, Histograms)
+        # define phase symmetries, currently only works for one phase
+        phase = G2stIO.GetPhaseNames(GPXfile)
+        phaseData = {}
+        for name in phase:
+            phaseData[name] = G2stIO.GetAllPhaseData(GPXfile, name)
+            symmetry[name] = phaseData[name]['General']['SGData']['SGSys']
 
         # Save the instance parameters
         self._Histograms = Histograms
@@ -115,6 +122,8 @@ class Calculator:
         self._restraintDict = restraintDict
         self._rigidbodyDict = rigidbodyDict
         self._rbIds = rbIds
+        self._symmetry = symmetry
+        self._phase = phase
         try:
             self._lowerLimit = Histograms[
                     list(Histograms.keys())[0]][u'Limits'][1][0]
@@ -153,112 +162,88 @@ class Calculator:
                 self._Phases, self._calcControls, self._pawleyLookup)[0]
         return yc1
 
-    def UpdateLattice(self, parmVarDict):
+    def UpdateLattice(self, parmVarDict, phase_index):
         '''
-        Update the lattice parameters in the current model
+        Update the lattice parameters in the current model for all phases
         '''
-        tmp = check_lattice_parameters(parmVarDict)
-        a, b, c, alpha, beta, gamma = tmp['lattice']
-        parmVarDict = tmp['parmVarDict']
 
 # Symmetry Options:
-# Cubic, Tetragonal, Hexagonal, Rhombohedral,
-# Orthorhombic, Monoclinic, Triclinic
+# cubic, tetragonal, hexagonal, rhombohedral, trigonal,
+# orthorhombic, monoclinic, triclinic
 
-        if self.Symmetry == 'Cubic':
+        if self._symmetry[self._phase[phase_index]] == 'cubic':
+            a = parmVarDict[str(phase_index) + '::a']
             b = a
             c = a
             alpha = 90.
             beta = 90.
             gamma = 90.
-            self._parmDict['a'] = a
-        elif self.Symmetry == 'Tetragonal':
-            if a is None:
-                a = self._parmDict['a']
-            elif c is None:
-                c = self._parmDict['c']
+
+        elif self._symmetry[self._phase[phase_index]] == 'tetragonal':
+            a = parmVarDict[str(phase_index) + '::a']
             b = a
+            c = parmVarDict[str(phase_index) + '::c']
             alpha = 90.
             beta = 90.
             gamma = 90.
-            self._parmDict['a'] = a
-            self._parmDict['c'] = c
 
-        elif self.Symmetry == 'Hexagonal':
-            if a is None:
-                a = self._parmDict['a']
-            elif c is None:
-                c = self._parmDict['c']
+        elif self._symmetry[self._phase[phase_index]] == 'hexagonal':
+            a = parmVarDict[str(phase_index) + '::a']
             b = a
+            c = parmVarDict[str(phase_index) + '::c']
             alpha = 90.
-            beta = 120.
-            gamma = 90.
-            self._parmDict['a'] = a
-            self._parmDict['c'] = c
-        elif self.Symmetry == 'Rhombohedral':
-            if a is None:
-                a = self._parmDict['a']
-            elif alpha is None:
-                alpha = self._parmDict['beta']
+            beta = 90.
+            gamma = 120.
+
+        elif self._symmetry[self._phase[phase_index]] == 'trigonal':
+            a = parmVarDict[str(phase_index) + '::a']
+            b = a
+            c = parmVarDict[str(phase_index) + '::c']
+            alpha = 90.
+            beta = 90.
+            gamma = 120.
+
+        elif self._symmetry[self._phase[phase_index]] == 'rhombohedral':
+            a = parmVarDict[str(phase_index) + '::a']
             b = a
             c = a
+            alpha = parmVarDict[str(phase_index) + '::alpha']
             beta = alpha
             gamma = alpha
-            self._parmDict['a'] = a
-            self._parmDict['alpha'] = alpha
-        elif self.Symmetry == 'Orthorhombic':
-            if a is None:
-                a = self._parmDict['a']
-            elif b is None:
-                b = self._parmDict['b']
-            elif c is None:
-                c = self._parmDict['c']
+
+        elif self._symmetry[self._phase[phase_index]] == 'orthorhombic':
+            a = parmVarDict[str(phase_index) + '::a']
+            b = parmVarDict[str(phase_index) + '::b']
+            c = parmVarDict[str(phase_index) + '::c']
             alpha = 90.
             beta = 90.
             gamma = 90.
-            self._parmDict['a'] = a
-            self._parmDict['b'] = b
-            self._parmDict['c'] = c
-        elif self.Symmetry == 'Monoclinic':
-            if a is None:
-                a = self._parmDict['a']
-            elif b is None:
-                b = self._parmDict['b']
-            elif c is None:
-                c = self._parmDict['c']
-            elif beta is None:
-                beta = self._parmDict['beta']
+
+        elif self._symmetry[self._phase[phase_index]] == 'monoclinic':
+            a = parmVarDict[str(phase_index) + '::a']
+            b = parmVarDict[str(phase_index) + '::b']
+            c = parmVarDict[str(phase_index) + '::c']
+            beta = parmVarDict[str(phase_index) + '::beta']
             alpha = 90.
             gamma = 90.
-            self._parmDict['a'] = a
-            self._parmDict['b'] = b
-            self._parmDict['c'] = c
-            self._parmDict['beta'] = beta
-        elif self.Symmetry == 'Triclinic':
-            if a is None:
-                a = self._parmDict['a']
-            elif b is None:
-                b = self._parmDict['b']
-            elif c is None:
-                c = self._parmDict['c']
-            elif alpha is None:
-                alpha = self._parmDict['alpha']
-            elif beta is None:
-                beta = self._parmDict['beta']
-            elif gamma is None:
-                gamma = self._parmDict['gamma']
-            self._parmDict['a'] = a
-            self._parmDict['b'] = b
-            self._parmDict['c'] = c
-            self._parmDict['alpha'] = alpha
-            self._parmDict['beta'] = beta
-            self._parmDict['gamma'] = gamma
 
-        [G, g] = G2latt.cell2Gmat([a, b, c, alpha, beta, gamma])[0]
+        elif self._symmetry[self._phase[phase_index]] == 'triclinic':
+            a = parmVarDict[str(phase_index) + '::a']
+            b = parmVarDict[str(phase_index) + '::b']
+            c = parmVarDict[str(phase_index) + '::c']
+            alpha = parmVarDict[str(phase_index) + '::alpha']
+            beta = parmVarDict[str(phase_index) + '::beta']
+            gamma = parmVarDict[str(phase_index) + '::gamma']
 
-        LatticeUpdate = {'0::A0': G[0, 0], '0::A1': G[1, 1], '0::A2': G[2, 2],
-                         '0::A3': G[0, 1], '0::A4': G[0, 2], '0::A5': G[1, 2]}
-        # print G
+        A = G2latt.cell2A([a, b, c, alpha, beta, gamma])
+
+        LatticeUpdate = {str(phase_index) + '::A0': A[0],
+                         str(phase_index) + '::A1': A[1],
+                         str(phase_index) + '::A2': A[2],
+                         str(phase_index) + '::A3': A[3],
+                         str(phase_index) + '::A4': A[4],
+                         str(phase_index) + '::A5': A[5]}
+
         self._parmDict.update(LatticeUpdate)
         return parmVarDict
 
@@ -266,10 +251,9 @@ class Calculator:
         '''
             Update parameters in the current model
         '''
-        for key in parmVarDict.keys():
-            if key in ['a', 'b', 'c', 'alpha', 'beta', 'gamma']:
-                parmVarDict = self.UpdateLattice(parmVarDict)
-                break
+        for phase_index in range(len(self._phase)):
+            if ((str(phase_index) + '::a') in parmVarDict.keys()) is True:
+                parmVarDict = self.UpdateLattice(parmVarDict, phase_index)
 
         if 'EXT' in parmVarDict.keys():
             for key in self._parmDict.keys():
@@ -280,28 +264,16 @@ class Calculator:
 
         self._parmDict.update(parmVarDict)
 
-
-def check_lattice_parameters(parmVarDict):
-    a, b, c, alpha, beta, gamma = None, None, None, None, None, None
-    keys = list(parmVarDict.keys()).copy()
-    for key in keys:
-        if key == 'a':
-            a = parmVarDict[key]
-            parmVarDict.pop(key, None)
-        elif key == 'b':
-            b = parmVarDict[key]
-            parmVarDict.pop(key, None)
-        elif key == 'c':
-            c = parmVarDict[key]
-            parmVarDict.pop(key, None)
-        elif key == 'alpha':
-            alpha = parmVarDict[key]
-            parmVarDict.pop(key, None)
-        elif key == 'beta':
-            beta = parmVarDict[key]
-            parmVarDict.pop(key, None)
-        elif key == 'gamma':
-            gamma = parmVarDict[key]
-            parmVarDict.pop(key, None)
-    return dict(lattice=(a, b, c, alpha, beta, gamma),
-                parmVarDict=parmVarDict)
+    def convert_lattice(self, paramList, params):
+        '''
+            Convert between A and unit cell parameters. A is defined in GSAS-II
+            as A = [G11, G22, G33, 2*G12, 2*G13, 2*G23] with G as the
+            reciprocal metric tensor elements.
+        '''
+        latparamList = ['0::A0', '1::A0', '2::A0', '3::A0', '4::A0', '5::A0',
+                        '6::A0', '7::A0', '8::A0', '9::A0', '10::A0']
+        check = any(item in paramList for item in latparamList)
+        if check is True:
+            return G2latt.A2cell(params)
+        else:
+            return G2latt.cell2A(params)